Refactor: utils 폴더 정리
