+++ src/utils/convertDate.ts
@@ -1,4 +1,4 @@
-const convertDate = (value: number): string => {
+export const convertDate = (value: number): string => {
   const date = new Date(value);
   const time = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(
     2,
@@ -6,5 +6,3 @@ const convertDate = (value: number): string => {
   )}-${String(date.getDate()).padStart(2, '0')}`;
   return time;
 };
-
-export default convertDate;
