+++ src/utils/formatNumber.ts
@@ -1,5 +1,3 @@
-const formatNumber = (value: number): string => {
+export const formatNumber = (value: number): string => {
   return new Intl.NumberFormat().format(value);
 };
-
-export default formatNumber;
