+++ src/utils/getTodayDate.ts
@@ -1,4 +1,4 @@
-const getTodayDate = () => {
+export const getTodayDate = () => {
   const todayDate = new Date();
   const year = todayDate.getFullYear();
   const month = (todayDate.getMonth() + 1).toString().padStart(2, '0');
@@ -6,5 +6,3 @@ const getTodayDate = () => {
   const dateStr = year + '-' + month + '-' + date;
   return dateStr;
 };
-
-export default getTodayDate;
